svn merge -r 2366:2559 http://code.taobao.org/svn/tfs/branches/stable-2.5/
git-svn-id: http://code.taobao.org/svn/tfs@2560 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ trunk/src/nameserver/ns_define.cpp
@@ -226,7 +226,7 @@ namespace tfs
 
     void print_int64(const common::ArrayHelper<uint64_t>& servers, std::string& result)
     {
-      for (int32_t i = 0; i < servers.get_array_index(); ++i)
+      for (int64_t i = 0; i < servers.get_array_index(); ++i)
       {
         result += "/";
         result += tbsys::CNetUtil::addrToString(*servers.at(i));
